// app/(main)/(sidebar)/(discover)/category/type.ts
export type ListCategoryType = {
  iuProduct: number;
  url: string;
  createdAt: Date;
};

export type CTypeCategoryDescriptionTypes = {
  tarIuProduct: number;
  description: string;
  url: string;
  hashtag: string[];
  category: string[];
  totalLike: number;
  totalDislike: number;
  createdAt: Date;
};